+++ template-sample-app/template/src/components/views/stores/StoresDetail.js
@@ -1,19 +1,17 @@
 import React, { useState, useEffect, useMemo } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { NavLink, useLocation, useNavigate, useParams } from 'react-router-dom';
+import { NavLink, useLocation, useParams } from 'react-router-dom';
 
 import { makeStyles } from '@material-ui/core/styles';
 import {
   Breadcrumbs,
   CircularProgress,
   Divider,
-  IconButton,
   Grid,
   Typography,
   Link,
 } from '@material-ui/core';
 import NavigateNextIcon from '@material-ui/icons/NavigateNext';
-import CloseIcon from '@material-ui/icons/Close';
 
 import { WrapperWidgetUI, FormulaWidgetUI, HistogramWidgetUI } from '@carto/react/ui';
 import { updateLayer, selectSourceById, setViewState } from '@carto/react/redux';
@@ -22,7 +20,7 @@ import { getStore, getRevenuePerMonth } from 'models/StoreModel';
 import { LAYER_ID, MONTHS_LABELS } from './constants';
 import { Isochrone } from 'components/common/Isochrone';
 import { currencyFormatter } from 'utils/formatter';
-import { setError } from 'config/appSlice';
+import { setBottomSheetOpen, setError } from 'config/appSlice';
 
 export default function StoresDetail() {
   const [storeDetail, setStoreDetail] = useState(null);
@@ -31,7 +29,6 @@ export default function StoresDetail() {
   const { id } = useParams();
   const source = useSelector((state) => selectSourceById(state, 'storesSource'));
   const location = useLocation();
-  const navigate = useNavigate();
 
   const classes = useStyles();
 
@@ -76,6 +73,8 @@ export default function StoresDetail() {
 
     // Set selected store on the layer
     dispatch(updateLayer(LAYER_ID, { selectedStore: id }));
+    
+    dispatch(setBottomSheetOpen(true));
 
     return () => {
       dispatch(updateLayer(LAYER_ID, { selectedStore: null }));
@@ -101,10 +100,6 @@ export default function StoresDetail() {
 
   return (
     <div>
-      <IconButton onClick={() => navigate('/stores')} className={classes.closeDetail}>
-        <CloseIcon />
-      </IconButton>
-
       <div className={classes.storeDetail}>
         <Breadcrumbs
           separator={<NavigateNextIcon />}
@@ -116,7 +111,7 @@ export default function StoresDetail() {
           </Link>
           <Typography color='textPrimary'>Store detail</Typography>
         </Breadcrumbs>
-        <Typography variant='h5' gutterBottom>
+        <Typography variant='h5' gutterBottom className={classes.title}>
           {storeName(storeDetail)}
         </Typography>
         <Isochrone latLong={storeLatLong}></Isochrone>
@@ -144,19 +139,23 @@ export default function StoresDetail() {
           onError={onRevenuePerMonthWidgetError}
         ></HistogramWidgetUI>
       </WrapperWidgetUI>
+
+      <Divider />
     </div>
   );
 }
 
 const useStyles = makeStyles((theme) => ({
-  closeDetail: {
-    position: 'absolute',
-    top: theme.spacing(1.75),
-    right: theme.spacing(1.75),
-    color: theme.palette.primary.main,
-  },
   storeDetail: {
     padding: theme.spacing(3.25, 3),
+
+    [theme.breakpoints.down('sm')]: {
+      paddingTop: theme.spacing(2),
+    },
+  },
+  title: {
+    textTransform: 'capitalize',
+    marginBottom: 4,
   },
   isochrone: {
     width: '100%',
@@ -174,5 +173,5 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 function storeName(store) {
-  return `${store.address}, ${store.city}`;
+  return `${store.address}, ${store.city}`.toLowerCase();
 }