+++ template-sample-app/template/src/config/appSlice.js
@@ -5,6 +5,7 @@ const slice = createSlice({
   initialState: {
     error: null,
     isolineResult: null,
+    bottomSheetOpen: false,
   },
   reducers: {
     setIsolineResult: (state, action) => {
@@ -13,6 +14,9 @@ const slice = createSlice({
     setError: (state, action) => {
       state.error = action.payload;
     },
+    setBottomSheetOpen: (state, action) => {
+      state.bottomSheetOpen = action.payload;
+    },
   },
 });
 
@@ -20,3 +24,7 @@ export default slice.reducer;
 
 export const setIsolineResult = (payload) => ({ type: 'app/setIsolineResult', payload });
 export const setError = (payload) => ({ type: 'app/setError', payload });
+export const setBottomSheetOpen = (payload) => ({
+  type: 'app/setBottomSheetOpen',
+  payload,
+});