+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## Not released
+- Fix selected store state in sample-app [#142](https://github.com/CartoDB/carto-react-template/pull/142)
+
 ## 1.0.0-beta5 (2020-11-27)
 - Adapt layout for mobile [#134](https://github.com/CartoDB/carto-react-template/pull/134)
 - Update favicons for better visibility on dark themes [#136](https://github.com/CartoDB/carto-react-template/pull/136)