+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Not released
+## 1.0.0-beta6 (2020-12-04)
 - Improve layout for mobile (specially for iOS) [#141](https://github.com/CartoDB/carto-react-template/pull/141)
 - Fix selected store state in sample-app [#142](https://github.com/CartoDB/carto-react-template/pull/142)